Address cubic/CodeRabbit tail review of ec5005e
- efcard_access: prefer a *supported* PACEInfo when EF.CardAccess holds
  multiple SecurityInfos (both bots), instead of taking the first parsed.
- diffie_hellman: reject a degenerate mapped generator (G' <= 1).
- iso7816/sm: do99 takes u16 so the status word is always exactly 2 bytes
  (rejects out-of-range statuses at the type level).
- docs: tlv decode_length lists IndefiniteLength; command_apdu field doc
  corrects the extended Le=256 (01 00) encoding; efdg11 test comment.

Deferred (reported): public_key_pace over-wide-coordinate rejection
(unreachable invariant, would force to_bytes->Result) and dba_key
unwrap_or(0) seed GIGO (MRZ is validated at the check-digit layer).

cargo clean + build: 0 warnings; 615 lib + 34 doctests pass.

+++ src/proto/iso7816/sm.rs
@@ -82,7 +82,9 @@ pub fn do97(ne: u32) -> Vec<u8> {
 /// The status word is always exactly 2 bytes. A minimal-integer encoding would
 /// drop leading zero bytes (e.g. a status with `SW1 == 0x00`), emitting a
 /// 1-byte or empty DO'99'; the SW must be carried verbatim as two bytes.
-pub fn do99(sw: u32) -> Vec<u8> {
+pub fn do99(sw: u16) -> Vec<u8> {
+    // `sw` is a u16, so the status word always encodes as exactly two bytes —
+    // out-of-range statuses are rejected at the type level.
     build_do(TAG_DO99, &int_to_bin(sw as u64, 2))
 }
